refactor(MessageContent): factor out EditMessage, types, Container to separate files, rename Content -> Markdown
