+++ client/src/components/Messages/Content/Container.tsx
@@ -1,0 +1,6 @@
+// Container Component
+const Container = ({ children }: { children: React.ReactNode }) => (
+  <div className="flex min-h-[20px] flex-grow flex-col items-start gap-4">{children}</div>
+);
+
+export default Container;
