+++ client/src/components/Messages/Content/MessageContent.tsx
@@ -1,44 +1,11 @@
-import { useRef, Fragment } from 'react';
-import { useRecoilState } from 'recoil';
-import { useUpdateMessageMutation } from 'librechat-data-provider';
-import type { TMessage, TResPlugin } from 'librechat-data-provider';
-import type { TAskFunction } from '~/common';
+import { Fragment } from 'react';
+import type { TResPlugin } from 'librechat-data-provider';
+import type { TMessageContent, TText, TDisplayProps } from '~/common';
 import { cn, getError } from '~/utils';
-import store from '~/store';
-import Content from './Content';
+import EditMessage from './EditMessage';
+import Container from './Container';
+import Markdown from './Markdown';
 import Plugin from './Plugin';
-
-type TInitialProps = {
-  text: string;
-  edit: boolean;
-  error: boolean;
-  unfinished: boolean;
-  isSubmitting: boolean;
-  isLast: boolean;
-};
-type TAdditionalProps = {
-  ask: TAskFunction;
-  message: TMessage;
-  isCreatedByUser: boolean;
-  siblingIdx: number;
-  enterEdit: (cancel: boolean) => void;
-  setSiblingIdx: (value: number) => void;
-};
-
-type TMessageContent = TInitialProps & TAdditionalProps;
-
-type TText = Pick<TInitialProps, 'text'>;
-type TEditProps = Pick<TInitialProps, 'text' | 'isSubmitting'> &
-  Omit<TAdditionalProps, 'isCreatedByUser'>;
-type TDisplayProps = TText &
-  Pick<TAdditionalProps, 'isCreatedByUser' | 'message'> & {
-    showCursor?: boolean;
-  };
-
-// Container Component
-const Container = ({ children }: { children: React.ReactNode }) => (
-  <div className="flex min-h-[20px] flex-grow flex-col items-start gap-4">{children}</div>
-);
 
 // Error Message Component
 const ErrorMessage = ({ text }: TText) => (
@@ -48,110 +15,6 @@ const ErrorMessage = ({ text }: TText) => (
     </div>
   </Container>
 );
-
-// Edit Message Component
-const EditMessage = ({
-  text,
-  message,
-  isSubmitting,
-  ask,
-  enterEdit,
-  siblingIdx,
-  setSiblingIdx,
-}: TEditProps) => {
-  const [messages, setMessages] = useRecoilState(store.messages);
-  const textEditor = useRef<HTMLDivElement | null>(null);
-  const { conversationId, parentMessageId, messageId } = message;
-  const updateMessageMutation = useUpdateMessageMutation(conversationId ?? '');
-
-  const resubmitMessage = () => {
-    const text = textEditor?.current?.innerText ?? '';
-    if (message.isCreatedByUser) {
-      ask({
-        text,
-        parentMessageId,
-        conversationId,
-      });
-
-      setSiblingIdx((siblingIdx ?? 0) - 1);
-    } else {
-      const parentMessage = messages?.find((msg) => msg.messageId === parentMessageId);
-
-      if (!parentMessage) {
-        return;
-      }
-      ask(
-        { ...parentMessage },
-        {
-          editedText: text,
-          editedMessageId: messageId,
-          isRegenerate: true,
-          isEdited: true,
-        },
-      );
-
-      setSiblingIdx((siblingIdx ?? 0) - 1);
-    }
-
-    enterEdit(true);
-  };
-
-  const updateMessage = () => {
-    if (!messages) {
-      return;
-    }
-    const text = textEditor?.current?.innerText ?? '';
-    updateMessageMutation.mutate({
-      conversationId: conversationId ?? '',
-      messageId,
-      text,
-    });
-    setMessages(() =>
-      messages.map((msg) =>
-        msg.messageId === messageId
-          ? {
-            ...msg,
-            text,
-          }
-          : msg,
-      ),
-    );
-    enterEdit(true);
-  };
-
-  return (
-    <Container>
-      <div
-        data-testid="message-text-editor"
-        className="markdown prose dark:prose-invert light w-full whitespace-pre-wrap break-words border-none focus:outline-none"
-        contentEditable={true}
-        ref={textEditor}
-        suppressContentEditableWarning={true}
-      >
-        {text}
-      </div>
-      <div className="mt-2 flex w-full justify-center text-center">
-        <button
-          className="btn btn-primary relative mr-2"
-          disabled={isSubmitting}
-          onClick={resubmitMessage}
-        >
-          Save & Submit
-        </button>
-        <button
-          className="btn btn-secondary relative mr-2"
-          disabled={isSubmitting}
-          onClick={updateMessage}
-        >
-          Save
-        </button>
-        <button className="btn btn-neutral relative" onClick={() => enterEdit(true)}>
-          Cancel
-        </button>
-      </div>
-    </Container>
-  );
-};
 
 // Display Message Component
 const DisplayMessage = ({ text, isCreatedByUser, message, showCursor }: TDisplayProps) => (
@@ -163,7 +26,7 @@ const DisplayMessage = ({ text, isCreatedByUser, message, showCursor }: TDisplay
       )}
     >
       {!isCreatedByUser ? (
-        <Content content={text} message={message} showCursor={showCursor} />
+        <Markdown content={text} message={message} showCursor={showCursor} />
       ) : (
         <>{text}</>
       )}
